moved pathfinder to own file
